fix: move to Upload.jsx function to download pdf
